Move auth to its own file
